refactor: User -> Search 수정
