+++ SearchService/src/main/resources/application.properties
@@ -1,6 +1,6 @@
-spring.application.name=UserService
+spring.application.name=SearchService
 
-server.port = 8081
+server.port = 8088
 
 # DB connect
 spring.datasource.url=${SPRING_DATASOURCE_URL}
