+++ SearchService/src/test/java/ready_to_marry/searchservice/SearchServiceApplicationTests.java
@@ -1,10 +1,10 @@
-package ready_to_marry.userservice;
+package ready_to_marry.searchservice;
 
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
 @SpringBootTest
-class UserServiceApplicationTests {
+class SearchServiceApplicationTests {
 
     @Test
     void contextLoads() {
